+++ app/Http/Controllers/MessageController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Comment;
 use App\Contact as Contact;
 use App\FirewallIp;
+use App\Repository\FireWallRepository;
 use App\Tags as Tags;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Crypt;
@@ -26,6 +27,10 @@ class MessageController extends Controller
 
         $contact = Contact::where( 'id', $id )->first();
 
+        if (!$contact){
+            return redirect()->route('admin_msg');
+        }
+
         $contact->delete();
 
         $request->session()->flash('message', "Enregistrer avec succès");
@@ -35,6 +40,11 @@ class MessageController extends Controller
     public function show(Request $request, $id)
     {
         $contact = Contact::where( 'id', $id )->first();
+
+        if (!$contact){
+            return redirect()->route('admin_msg');
+        }
+
         return view('admin.viewLivewire', [
             "model" => $contact,
         ]);
@@ -44,23 +54,17 @@ class MessageController extends Controller
     {
 
         $contact = Contact::where( 'id', $id )->first();
+
+        if (!$contact){
+            return redirect()->route('admin_msg');
+        }
+
         if(!empty(trim($contact->ip))) {
-            FirewallIp::firstOrCreate([
-                "ip" => $contact->ip
-            ]);
+            FireWallRepository::createReport($contact->ip, 2, "Bannie par l'administrateur");
         }
 
         $request->session()->flash('message', "ban");
         return redirect()->route('admin_msg');
     }
 
-    public function deleteAll(Request $request) {
-        DB::statement('PRAGMA foreign_keys = OFF;');
-
-        $request->session()->flash('message', "Supprimé avec succès");
-        DB::table("contact")->delete();
-        DB::statement('PRAGMA foreign_keys = ON;');
-
-        return redirect()->route('admin_comment');
-    }
 }
